refactor: Restructure files and packages
