+++ api/cmd/api/http_test.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/kava-labs/kavachat/api/cmd/api/config"
+	"github.com/kava-labs/kavachat/api/internal/config"
 )
 
 const (
