+++ api/internal/middleware/model_allowlist_test.go
@@ -8,8 +8,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/kava-labs/kavachat/api/cmd/api/config"
-	"github.com/kava-labs/kavachat/api/cmd/api/types"
+	"github.com/kava-labs/kavachat/api/internal/config"
+	"github.com/kava-labs/kavachat/api/internal/types"
 	"github.com/openai/openai-go"
 	"github.com/rs/zerolog/log"
 	"github.com/stretchr/testify/require"
