+++ api/internal/middleware/model_extract.go
@@ -8,7 +8,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/kava-labs/kavachat/api/cmd/api/types"
+	"github.com/kava-labs/kavachat/api/internal/types"
 	"github.com/openai/openai-go"
 	"github.com/rs/zerolog"
 )
